tools/python/requirements.txt
qrcode==7.4.2
Pillow==10.2.0
numpy==1.26.4 